Fixed a ton of files and made a new player
The files are cleaner, no more warning, and I am making a new player with all new animations and fixing the tools that he has. I made a bunch of small fixed and hopefully by tomorrow I will be in a state to actually fix the mechanics rather than the player.