Artwork and Tutorial Level Start
I have added the first round of tiles and then I began to design level 1 of the game. I want to have a finished level by the end of the month, something that I can show off and play and use as a demo for friends and family. Also for play testing and getting feedback from people